+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - [#60] Accessing the global "loadTemplates" which is now namespaced under foundry.applications.handlebars.loadTemplates
 - [#61] Accessing the global "Token" which is now namespaced under foundry.canvas.placeables.Token
 - [#62] Accessing the global "Hotbar" which is now namespaced under foundry.applications.ui.Hotbar
+- [#63] Accessing the global "TextEditor" which is now namespaced under foundry.applications.ux.TextEditor.implementation
 
 ---